+++ packages/desktop/src/menu/MenuItem.tsx
@@ -1,8 +1,6 @@
 import React, {forwardRef, ReactNode} from 'react'
-
-import {Flex, FlexItem, Icon, Box, Section, Stub} from '@qiwi/pijma-core'
-
-import {Paragraph} from '../typography'
+import {Flex, FlexItem, Icon, Box, Section, Stub, Image} from '@qiwi/pijma-core'
+import {Paragraph, Text} from '../typography'
 
 export interface MenuItemProps {
   text: string
@@ -13,19 +11,26 @@ export interface MenuItemProps {
   hover?: boolean
   active?: boolean
   focus?: boolean
+  size?: 's' | 'm'
   stub?: boolean
+}
+
+const IconSize: Record<NonNullable<MenuItemProps['size']>, number> = {
+  s: 6,
+  m: 12,
 }
 
 export const MenuItem = forwardRef<HTMLDivElement, MenuItemProps>(({
   text,
   notes,
   icon,
-  stub,
   submenu = false,
   round = false,
   hover = false,
   active = false,
   focus = false,
+  size = 's',
+  stub = false,
   ...props
 }, ref) => (
   <Section
@@ -40,11 +45,19 @@ export const MenuItem = forwardRef<HTMLDivElement, MenuItemProps>(({
       {icon ? (
         stub ? (
           <FlexItem align={notes ? undefined : 'center'} shrink={0} mr={4}>
-            <Stub r={12} width={6} height={6}/>
+            <Stub r={IconSize[size] * 2} width={IconSize[size]} height={IconSize[size]}/>
           </FlexItem>
         ) : (
           <FlexItem align={notes ? undefined : 'center'} shrink={0} mr={4}>
-            {icon}
+            {typeof icon === 'string' ? (
+              <Image
+                src={icon}
+                width={IconSize[size]}
+                height={IconSize[size]}
+              />
+            ) : (
+              icon
+            )}
           </FlexItem>
         )
       ) : (
@@ -53,16 +66,16 @@ export const MenuItem = forwardRef<HTMLDivElement, MenuItemProps>(({
       <FlexItem align="center" grow={1}>
         <Flex justify="center" direction="column">
           {stub ? (
-            <Box width={80} maxWidth={1}>
-              <Paragraph stub clamp={icon ? 1 : undefined}/>
+            <Box maxWidth={38} width={1}>
+              <Text display="block" size="m" stub/>
             </Box>
           ) : (
             <Paragraph clamp={icon && !notes ? 2 : undefined} bold>{text}</Paragraph>
           )}
           {notes ? (
             stub ? (
-              <Box mt={1} width={100} maxWidth={1}>
-                <Paragraph size="s" stub clamp={1}/>
+              <Box mt={1} width={1} maxWidth={18}>
+                <Text display="block" size="s" stub/>
               </Box>
             ) : (
               <Box mt={1}>
@@ -77,15 +90,9 @@ export const MenuItem = forwardRef<HTMLDivElement, MenuItemProps>(({
         </Flex>
       </FlexItem>
       {submenu ? (
-        stub ? (
-          <FlexItem align="center" shrink={0} width={6} height={6} ml={3}>
-            <Stub width={6} height={6}/>
-          </FlexItem>
-        ) : (
-          <FlexItem align="center" shrink={0} width={6} height={6} ml={3}>
-            <Icon name="angle-right"/>
-          </FlexItem>
-        )
+        <FlexItem align="center" shrink={0} width={6} height={6} ml={3}>
+          <Icon name="angle-right"/>
+        </FlexItem>
       ) : (
         null
       )}