+++ packages/mobile/CHANGELOG.md
@@ -1,3 +1,80 @@
+# [@qiwi/pijma-mobile-v1.108.0](https://github.com/qiwi/pijma/compare/v1.107.1...v1.108.0) (2020-02-04)
+
+
+### Features
+
+* **mobile:** stub header menu ([efae3e1](https://github.com/qiwi/pijma/commit/efae3e1))
+
+# [@qiwi/pijma-mobile-v1.107.1](https://github.com/qiwi/pijma/compare/v1.107.0...v1.107.1) (2020-02-04)
+
+
+### Bug Fixes
+
+* rating ([b04a420](https://github.com/qiwi/pijma/commit/b04a420))
+
+# [@qiwi/pijma-mobile-v1.107.0](https://github.com/qiwi/pijma/compare/v1.106.0...v1.107.0) (2020-02-04)
+
+
+### Features
+
+* stub footer-app ([409e174](https://github.com/qiwi/pijma/commit/409e174))
+
+# [@qiwi/pijma-mobile-v1.105.0](https://github.com/qiwi/pijma/compare/v1.104.0...v1.105.0) (2020-02-03)
+
+
+### Features
+
+* stub menu-link ([2b41552](https://github.com/qiwi/pijma/commit/2b41552))
+
+# [@qiwi/pijma-mobile-v1.104.0](https://github.com/qiwi/pijma/compare/v1.103.0...v1.104.0) (2020-01-31)
+
+
+### Features
+
+* suggest-field ([945e2f9](https://github.com/qiwi/pijma/commit/945e2f9))
+
+# [@qiwi/pijma-mobile-v1.103.0](https://github.com/qiwi/pijma/compare/v1.102.0...v1.103.0) (2020-01-31)
+
+
+### Features
+
+* stub footer-out ([26b706a](https://github.com/qiwi/pijma/commit/26b706a))
+
+# [@qiwi/pijma-mobile-v1.102.0](https://github.com/qiwi/pijma/compare/v1.101.1...v1.102.0) (2020-01-31)
+
+
+### Features
+
+* stub footer-nav ([e3eb002](https://github.com/qiwi/pijma/commit/e3eb002))
+
+# [@qiwi/pijma-mobile-v1.101.1](https://github.com/qiwi/pijma/compare/v1.101.0...v1.101.1) (2020-01-31)
+
+
+### Bug Fixes
+
+* stub text-field ([caaddbb](https://github.com/qiwi/pijma/commit/caaddbb))
+
+# [@qiwi/pijma-mobile-v1.100.0](https://github.com/qiwi/pijma/compare/v1.99.0...v1.100.0) (2020-01-29)
+
+
+### Features
+
+* stub menu-item ([1fc97d2](https://github.com/qiwi/pijma/commit/1fc97d2))
+
+# [@qiwi/pijma-mobile-v1.99.0](https://github.com/qiwi/pijma/compare/v1.98.1...v1.99.0) (2020-01-29)
+
+
+### Features
+
+* stub logo-block-link ([c012a26](https://github.com/qiwi/pijma/commit/c012a26))
+
+# [@qiwi/pijma-mobile-v1.98.1](https://github.com/qiwi/pijma/compare/v1.98.0...v1.98.1) (2020-01-29)
+
+
+### Bug Fixes
+
+* export logo-block-link ([9a136b0](https://github.com/qiwi/pijma/commit/9a136b0))
+
 # [@qiwi/pijma-mobile-v1.98.0](https://github.com/qiwi/pijma/compare/v1.97.0...v1.98.0) (2020-01-23)